OS-51, OS-52: Feat: Add configurable log levels for libraries
This change introduces the ability to configure the log levels for
third-party libraries independently of the main application's log level.

This is useful for reducing log noise from verbose libraries (like
selenium and urllib3) when the application's log level is set to DEBUG.

A new `LIBRARY_LOG_LEVELS` dictionary has been added to the `Config`
class, which can be overridden by an environment variable. The
`setup_logging` function now applies these levels to the specified
library loggers.